refactor(otp-input-kit): move source files into entry subpackage
