+++ otp-input-kit/src/main/java/com/burkido/otpinputkit/entry/OtpInputState.kt
@@ -1,4 +1,4 @@
-package com.burkido.otpinputkit
+package com.burkido.otpinputkit.entry
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.Stable
