+++ otp-input-kit/src/main/java/com/burkido/otpinputkit/entry/cell/CursorIndicator.kt
@@ -1,4 +1,4 @@
-package com.burkido.otpinputkit.cell
+package com.burkido.otpinputkit.entry.cell
 
 import androidx.compose.animation.core.RepeatMode
 import androidx.compose.animation.core.animateFloat
@@ -14,7 +14,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.alpha
-import com.burkido.otpinputkit.style.OtpCursorConfig
+import com.burkido.otpinputkit.entry.style.OtpCursorConfig
 
 /**
  * Internal blinking cursor indicator for focused OTP cells.
