+++ otp-input-kit/src/main/java/com/burkido/otpinputkit/entry/cell/UnderlineOtpCell.kt
@@ -1,4 +1,4 @@
-package com.burkido.otpinputkit.cell
+package com.burkido.otpinputkit.entry.cell
 
 import androidx.compose.animation.AnimatedContent
 import androidx.compose.animation.EnterTransition
@@ -31,15 +31,15 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.dp
-import com.burkido.otpinputkit.OtpInputDefaults
-import com.burkido.otpinputkit.style.CellAnimation
-import com.burkido.otpinputkit.style.OtpAnimationSpec
-import com.burkido.otpinputkit.style.OtpCellColors
-import com.burkido.otpinputkit.style.OtpCellDimensions
-import com.burkido.otpinputkit.style.OtpCellPadding
-import com.burkido.otpinputkit.style.OtpCursorConfig
-import com.burkido.otpinputkit.style.OtpMaskConfig
-import com.burkido.otpinputkit.style.OtpTextStyles
+import com.burkido.otpinputkit.entry.OtpInputDefaults
+import com.burkido.otpinputkit.entry.style.CellAnimation
+import com.burkido.otpinputkit.entry.style.OtpAnimationSpec
+import com.burkido.otpinputkit.entry.style.OtpCellColors
+import com.burkido.otpinputkit.entry.style.OtpCellDimensions
+import com.burkido.otpinputkit.entry.style.OtpCellPadding
+import com.burkido.otpinputkit.entry.style.OtpCursorConfig
+import com.burkido.otpinputkit.entry.style.OtpMaskConfig
+import com.burkido.otpinputkit.entry.style.OtpTextStyles
 import kotlinx.coroutines.delay
 import kotlin.math.roundToInt
